chore: Refactoring components/CryptoCurrencyIcon
