--- web/src/components/CryptoCurrencyIcon/CryptoCurrencyIcon.postcss
@@ -1,14 +1,0 @@
-.iconSmall {
-  height: 24px;
-  width: 24px;
-}
-
-.iconMedium {
-  height: 36px;
-  width: 36px;
-}
-
-.iconLarge {
-  height: 64px;
-  width: 64px;
-}
